backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
Pillow==11.1.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
diffusers==0.31.0
accelerate==1.2.1
safetensors==0.4.5
xformers==0.0.28.post3
python-multipart==0.0.20
